requirements.txt
django
python-novaclient
python-keystoneclient
keystoneauth1
django-bootstrap3
django-phonenumber-field
uswgi
python-glanceclient
